Started refactor on the converter classes. Made certain methods static. Considering getting rid of the converter classes all together and moving the convert methods to either the Json classes or the Database/Gacha/Playbook classes
